// src/app/app_entities/dto/processo.dto.ts
export class DocumentosProcessoDTO {
    labelDocumentoPrimario?: string;
    tipoDocumentoPrimario?: string;
    nomeDocumentoPrimario?: string;
    linkDocumentoPrimario?: string;
    labelDocumentoSecundario?: string;
    tipoDocumentoSecundario?: string;
    nomeDocumentoSecundario?: string;
    linkDocumentoSecundario?: string;
  }
  